requirements.txt
apache-airflow=2.10.2